embedding/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.4.0
    # via
    #   httpx
    #   openai
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
boto3==1.29.1
    # via -r requirements.in
botocore==1.32.7
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
certifi==2024.7.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
distro==1.9.0
    # via openai
ecdsa==0.19.0
    # via python-jose
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jose==1.0.0
    # via -r requirements.in
jsonschema==4.22.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
numpy==2.1.0
    # via pgvector
openai==1.34.0
    # via -r requirements.in
pgvector==0.2.4
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.0
    # via
    #   python-jose
    #   rsa
pydantic==2.7.4
    # via
    #   -r requirements.in
    #   openai
pydantic-core==2.18.4
    # via pydantic
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.7.24
    # via tiktoken
requests==2.32.3
    # via
    #   -r requirements.in
    #   tiktoken
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via python-jose
s3transfer==0.7.0
    # via boto3
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tiktoken==0.5.2
    # via -r requirements.in
tqdm==4.66.5
    # via openai
typing-extensions==4.12.2
    # via
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.0.7
    # via
    #   botocore
    #   requests
